requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
beautifulsoup4==4.9.3
chardet==3.0.4
idna==3.1
lxml==4.6.2
multidict==5.1.0
PyYAML==5.4.1
soupsieve==2.1
typing-extensions==3.7.4.3
yarl==1.6.3
